Fixed Parser
When I changed the parser to only recognize leave and redo at the end
of blocks, I broke the parser, and quite haughtily uploaded without
testing.  It now works as intended, and break/continue as terms
have been replaced with leave/redo from the parser (but not the AST
tree).