bug fix, break omni proc
Added a break proc, that doesn't do anything.  But you can set
"break_on_me" in the debugger, and then call omni.break in the code,
and it will work just fine.

Also, fixed a problem with operator precedence, I had the logical
operators at the highest precedence, instead of the lowest.