New Features!
(*) map and sequence initializers.  Good for assignments and proc calls.

(*) move/copy operators as proc arguments.

(*) variable declarations can now include an initial assignment

Also, changed to use Berkeley Yacc (Bison compatibility will not be thwarted, but it will no longer be the primary target.)

I guess it doesn't look like much on paper, but it really changes the feel of the language.